Refactored _project method into base class Tracker.
